requirements.txt
chromadb>=0.5.0
sentence-transformers>=2.6.0
datasets>=2.18.0
openai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
pandas>=2.0.0
tqdm>=4.66.0